proxy

+++ docker/proxy/s6/w9init/migration.sh
@@ -3,14 +3,14 @@
 set +e 
 
 nginx_proxy(){
-    cp /etc/initproxy.conf /data/nginx/proxy_host/
-    # if [ ! -f /data/nginx/proxy_host/initproxy.conf ] || [ $(stat -c %Y /etc/shadow) -ge $(stat -c %Y /data/nginx/proxy_host/initproxy.conf) ]
-    # then
-    #     cp /etc/initproxy.conf /data/nginx/proxy_host/
-    #     echo "Update initproxy.conf to Nginx"
-    # else
-    #     echo "Don't need to update initproxy.conf to Nginx"
-    # fi
+
+    if [ ! -f /data/nginx/proxy_host/initproxy.conf ] || [ $(stat -c %Y /etc/shadow) -ge $(stat -c %Y /data/nginx/proxy_host) ]
+    then
+        cp /etc/initproxy.conf /data/nginx/proxy_host/
+        echo "Update initproxy.conf to Nginx"
+    else
+        echo "Don't need to update initproxy.conf to Nginx"
+    fi
     
 }
